requirements.txt
streamlit
pandas
mysql-connector-python
python-dateutil
matplotlib